# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DataFlow API
  description: Secure, self-hosted family data vault API
  version: 0.2.0
  license:
    name: MIT
servers:
  - url: http://localhost:3460
    description: Development server
paths:
...
  /api/categories/{id}/share:
    post:
      summary: Share a category
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category shared
  /api/categories/{id}/share/{shareUserId}:
    delete:
      summary: Unshare a category
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: shareUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category unshared
  /api/categories/{id}/shares:
    get:
      summary: List category shares
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of shares
...
  /api/items/{id}/share:
    post:
      summary: Share an item
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item shared
  /api/items/{id}/share/{shareUserId}:
    delete:
      summary: Unshare an item
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: shareUserId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item unshared
  /api/items/{id}/shares:
    get:
      summary: List item shares
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of shares
...
  /api/members/{id}:
    get:
      summary: Get member details
      tags: [Members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Member details
    put:
      summary: Update member
      tags: [Members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Member updated
    delete:
      summary: Delete member (admin only)
      tags: [Members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Member deleted
  /api/members/{id}/deactivate:
    put:
      summary: Deactivate member (admin only)
      tags: [Members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Member deactivated
  /api/members/{id}/activate:
    put:
      summary: Activate member (admin only)
      tags: [Members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Member activated
  /api/members/{id}/unlock:
    post:
      summary: Unlock member (admin only)
      tags: [Members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Member unlocked
...
  /api/security/breach-check/{prefix}:
    get:
      summary: Check password breach (k-anonymity)
      tags: [Security]
      parameters:
        - name: prefix
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Breach check result
...
  /api/security/totp/generate/{itemId}/{fieldId}:
    get:
      summary: Generate TOTP code
      tags: [Security]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
        - name: fieldId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: TOTP code
...